Migrate files to place UI components that are specific to NeovimEditor nearby
