+++ browser/src/Editor/NeovimEditor/NeovimEditor.tsx
@@ -45,10 +45,10 @@ import { workspace } from "./../../Services/Workspace"
 import { Editor, IEditor } from "./../Editor"
 
 import { BufferManager } from "./../BufferManager"
-import { CompletionMenu } from "./../CompletionMenu"
-import { HoverRenderer } from "./../HoverRenderer"
-import { NeovimPopupMenu } from "./../NeovimPopupMenu"
-import { NeovimSurface } from "./../NeovimSurface"
+import { CompletionMenu } from "./CompletionMenu"
+import { HoverRenderer } from "./HoverRenderer"
+import { NeovimPopupMenu } from "./NeovimPopupMenu"
+import { NeovimSurface } from "./NeovimSurface"
 
 import { ContextMenuManager } from "./../../Services/ContextMenu"
