+++ browser/src/Services/Sidebar/SidebarView.tsx
@@ -9,7 +9,7 @@ import { connect } from "react-redux"
 
 import { IEvent } from "oni-types"
 
-import { KeyboardInputView } from "./../../Editor/KeyboardInput"
+import { KeyboardInputView } from "./../../Input/KeyboardInput"
 import { Icon, IconSize } from "./../../UI/Icon"
 
 import { ISidebarEntry, ISidebarState } from "./SidebarStore"
