+++ browser/src/UI/components/Tabs.tsx
@@ -9,7 +9,7 @@ import { connect } from "react-redux"
 
 import * as classNames from "classnames"
 
-import * as BufferSelectors from "./../selectors/BufferSelectors"
+import * as BufferSelectors from "./../../Editor/NeovimEditor/NeovimEditorSelectors"
 import * as State from "./../../Editor/NeovimEditor/NeovimEditorStore"
 
 import { addDefaultUnitIfNeeded } from "./../../Font"
